Restructured the whole project.
