+++ src/ui/home_screen.py
@@ -8,12 +8,12 @@ from kivy.uix.screenmanager import Screen
 class HomeScreen(Screen):
     def __init__(self, **kwargs):
         super(HomeScreen, self).__init__(**kwargs)
-        self.background = Image(source='images/background.jpg', allow_stretch=True, keep_ratio=False)
+        self.background = Image(source='src/images/background.jpg', allow_stretch=True, keep_ratio=False)
         self.add_widget(self.background)
 
         self.label = Label(
             text='RPG Helper', 
-            font_name='fonts/BLKCHCRY.TTF',
+            font_name='src/fonts/BLKCHCRY.TTF',
             font_size='40sp', 
             size_hint=(.25, .10), 
             pos_hint={'center_x' : .5, 'center_y' : .70})
@@ -26,19 +26,19 @@ class HomeScreen(Screen):
         self.buttons.orientation = 'vertical'
 
         self.join_button = Button(
-            font_name='fonts/Enchanted_Land.otf',
+            font_name='src/fonts/Enchanted_Land.otf',
             font_size=24,
             text='Join')
         self.buttons.add_widget(self.join_button)
 
         self.host_button = Button(
-            font_name='fonts/Enchanted_Land.otf',
+            font_name='src/fonts/Enchanted_Land.otf',
             font_size=24,
             text='Host')
         self.buttons.add_widget(self.host_button)
 
         self.character_button = Button(
-            font_name='fonts/Enchanted_Land.otf',
+            font_name='src/fonts/Enchanted_Land.otf',
             font_size=24,
             text='Characters')
         self.buttons.add_widget(self.character_button)
@@ -46,7 +46,7 @@ class HomeScreen(Screen):
         self.add_widget(self.buttons)
 
         self.quit_button = Button(
-            font_name='fonts/Enchanted_Land.otf',
+            font_name='src/fonts/Enchanted_Land.otf',
             text='Quit', 
             font_size=24,
             size_hint=(.10, .05),
