+++ src/ui/host_screen.py
@@ -10,14 +10,14 @@ from kivy.graphics import *
 class HostScreen(Screen):
     def __init__(self, **kwargs):
         super(HostScreen, self).__init__(**kwargs)
-        self.background = Image(source='images/background.jpg',
+        self.background = Image(source='src/images/background.jpg',
                                 allow_stretch=True,
                                 keep_ratio=False)
         self.add_widget(self.background)
 
         # Back Button
         self.back_button = Button(text='Back',
-                                  font_name='fonts/vinque.ttf',
+                                  font_name='src/fonts/vinque.ttf',
                                   size_hint=(.10, .05),
                                   pos_hint={'center_x': .95, 'bottom_y': .025})
         self.add_widget(self.back_button)
@@ -28,7 +28,7 @@ class HostScreen(Screen):
 
         # A label displaying the page name
         self.label_host_game = Label(text="Host ye olde Game!",
-                                     font_name='fonts/BLKCHCRY.TTF',
+                                     font_name='src/fonts/BLKCHCRY.TTF',
                                      font_size=40,
                                      size_hint=(.25, .10),
                                      pos_hint={'center_x': .5, 'center_y': .8})
@@ -66,7 +66,7 @@ class HostScreen(Screen):
 
         # The button to start a game
         self.start_game_button = Button(text='Start Game',
-                                        font_name='fonts/vinque.ttf',
+                                        font_name='src/fonts/vinque.ttf',
                                         font_size=24,
                                         size_hint=(.4, .05),
                                         pos_hint={'center_x': .5, 'center_y': .5})
